+++ staramr/blast/results/pointfinder/BlastResultsParserPointfinder.py
@@ -68,7 +68,7 @@ class BlastResultsParserPointfinder(BlastResultsParser):
                 hit.get_genome_contig_start(),
                 hit.get_genome_contig_end(),
                 db_mutation.get_pointfinder_mutation_string(),
-                self._blast_database.get_notes(hit.get_amr_gene_id(), db_mutation)
+                self._blast_database.get_cge_notes(hit.get_amr_gene_id(), db_mutation)
                 ]
 
         return result