+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 * Fixed an issue where the string "None" in the drug table would be parsed differently by different versions of pandas (#175).
 * Upgraded to pandas version 2.
-* Added CGE-predicted phenotypes to Pointfinder output.
+* Added the CGE-predicted phenotypes, PMID, Mechanism, and the Required Mutation columns to the Pointfinder output.
 * The resfinder.tsv and pointfinder.tsv outputs now contain a Notes column.
 * Updated the help description of the --mlst-scheme parameter to include a more useful link for available schemas.
 * Switched to only officially supporting Python 3.7+ due to recent incompatibilities with Python 3.6 and some Python packages (numpy, biopython, and others).